requirements.txt
torch
opencv-python
mediapipe
pandas
numpy
ffmpeg-python
tqdm
